+++ packages/terra-tabs/tests/wdio/tab-pane-spec.js
@@ -3,45 +3,31 @@ describe('TabPane', () => {
 
   describe('Text Only', () => {
     before(() => browser.url('/#/raw/tests/terra-tabs/tabs/tab-pane/text-only-tab-pane'));
-    Terra.should.matchScreenshot();
-    Terra.should.beAccessible();
-    Terra.should.themeCombinationOfCustomProperties({
-      testName: 'themed',
-      properties: {
-        '--terra-tabs-font-size': '30px',
-        '--terra-tabs-line-height': '2',
-        '--terra-tabs-text-only-padding': '50px',
-      },
-    });
+    Terra.it.matchesScreenshot();
+    Terra.it.isAccessible();
   });
 
   describe('Icon Only', () => {
     before(() => browser.url('/#/raw/tests/terra-tabs/tabs/tab-pane/icon-only-tab-pane'));
-    Terra.should.matchScreenshot();
-    Terra.should.beAccessible();
-    Terra.should.themeEachCustomProperty({
-      '--terra-tabs-icon-only-padding': '50px',
-    });
+    Terra.it.matchesScreenshot();
+    Terra.it.isAccessible();
   });
 
   describe('Icon and Text', () => {
     before(() => browser.url('/#/raw/tests/terra-tabs/tabs/tab-pane/icon-and-text-tab-pane'));
-    Terra.should.matchScreenshot();
-    Terra.should.beAccessible();
-    Terra.should.themeEachCustomProperty({
-      '--terra-tabs-padding': '50px',
-    });
+    Terra.it.matchesScreenshot();
+    Terra.it.isAccessible();
   });
 
   describe('Custom Display', () => {
     before(() => browser.url('/#/raw/tests/terra-tabs/tabs/tab-pane/custom-display-tab-pane'));
-    Terra.should.matchScreenshot();
-    Terra.should.beAccessible();
+    Terra.it.matchesScreenshot();
+    Terra.it.isAccessible();
   });
 
   describe('Long text', () => {
     before(() => browser.url('/#/raw/tests/terra-tabs/tabs/tab-pane/long-text-tab-pane'));
-    Terra.should.matchScreenshot({ selector: '#longText' });
-    Terra.should.beAccessible();
+    Terra.it.matchesScreenshot({ selector: '#longText' });
+    Terra.it.isAccessible();
   });
 });