+++ packages/terra-application-layout/CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
-* Removed all inline styles and implemented the same styles using external css.
+* Removed all inline styles from Dev-site components and implemented the same styles using external css.
 
 5.1.0 - (March 15, 2019)
 ------------------