+++ packages/terra-application-layout/CHANGELOG.md
@@ -3,8 +3,15 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
 
-5.1.0 - (March 15, 2019)
+5.2.0 - (May 21, 2019)
+------------------
+### Changed
+* Patch dependency update
+
+5.1.0 - (May 15, 2019)
 ------------------
 ### Changed
 * Minor dependency update